requirements.txt
blinker==1.6.2
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
